+++ lib/mtbuild/test_application_configuration.rb
@@ -23,6 +23,7 @@ module MTBuild
 
       desc "Build and run test application '#{@parent_project.project_name}' with configuration '#{@configuration_name}'"
       new_task = test_application_task @configuration_name => dependencies do |t|
+        @post_build.call if @post_build.respond_to? :call
         puts "built test application #{t.name}."
         sh "\"#{application_binaries.first}\""
         puts "ran test application #{t.name}."