+++ CHANGES.md
@@ -7,6 +7,7 @@
 
 * Fixed issue where executables and libraries are generated with colons in
   their names due to workspace prefix.
+* Added pre-build and post-build project tasks. 
 
 
 ## MTBuild 0.1.1 ##